Add plots and notebook example

+++ lavague-core/lavague/core/evaluator.py
@@ -19,9 +19,13 @@ import os
 from llama_index.core import QueryBundle
 from matplotlib.figure import Figure
 import re
-from lavague.core.utilities.format_utils import keep_assignments, return_assigned_variables
+from lavague.core.utilities.format_utils import (
+    keep_assignments,
+    return_assigned_variables,
+)
 import matplotlib.pyplot as plt
-import numpy as np
+import seaborn as sns
+
 
 def init_driver() -> WebDriver:
     # these imports are necessary as they will be pasted to the output
@@ -34,13 +38,15 @@ def init_driver() -> WebDriver:
     chrome_options = Options()
     chrome_options.add_argument("--headless")
     chrome_options.add_argument("--no-sandbox")
-    chrome_options.add_argument('--proxy-server=127.0.0.1:9999')
+    chrome_options.add_argument("--proxy-server=127.0.0.1:9999")
 
     return webdriver.Chrome(options=chrome_options)
+
 
 class SeleniumDriverForEval(SeleniumDriver):
     def check_visibility(self, xpath: str) -> bool:
         return True
+
 
 class Evaluator(ABC):
     @abstractmethod
@@ -55,7 +61,7 @@ class Evaluator(ABC):
         """Loads a specific HTML content into the browser."""
         file_path = f"tmp_{uuid.uuid4()}.html"
 
-        with open(file_path, 'w', encoding='utf-8') as file:
+        with open(file_path, "w", encoding="utf-8") as file:
             file.write(html)
 
         abs_file_path = os.path.abspath(file_path)
@@ -66,8 +72,13 @@ class Evaluator(ABC):
         os.remove(abs_file_path)
 
     def _extract_backend_node_ids(self, html_content):
-        soup = BeautifulSoup(html_content, 'html.parser')
-        return set([tag['backend_node_id'] for tag in soup.find_all(attrs={"backend_node_id": True})])
+        soup = BeautifulSoup(html_content, "html.parser")
+        return set(
+            [
+                tag["backend_node_id"]
+                for tag in soup.find_all(attrs={"backend_node_id": True})
+            ]
+        )
 
     def _intersection_score(self, set1, set2):
         intersection_set = set1 & set2
@@ -82,42 +93,59 @@ class Evaluator(ABC):
         recall, precision = self._intersection_score(ground_truth_ids, context_ids)
         return recall, precision
 
+
 class RetrieverEvaluator(Evaluator):
     def __init__(self):
         self.driver = SeleniumDriverForEval(get_selenium_driver=init_driver)
-    
-    def rephrase_dataset(self, dataset: pd.DataFrame, csv_out_name: str, llm: LLM = None) -> pd.DataFrame:
-        assert not os.path.isfile(csv_out_name)
+
+    def rephrase_dataset(
+        self, dataset: pd.DataFrame, csv_out_name: str, llm: LLM = None
+    ) -> pd.DataFrame:
+        if os.path.isfile(csv_out_name):
+            raise ValueError(f"{csv_out_name} already exists")
         if llm is None:
             llm = get_default_context().llm
         rephraser = Rephraser(llm)
         for i, row in tqdm(dataset.iterrows()):
             rephrase_list = rephraser.rephrase_query(row["query"])
-            dataset.at[i, "retriever_query"] = rephrase_list[0]['query']
-            dataset.at[i, "llm_query"] = rephrase_list[0]['action']
+            dataset.at[i, "retriever_query"] = rephrase_list[0]["query"]
+            dataset.at[i, "llm_query"] = rephrase_list[0]["action"]
         dataset.to_csv(csv_out_name)
         return dataset
 
-    def evaluate(self, retriever: BaseHtmlRetriever, rephrased_dataset: pd.DataFrame, csv_out_name: str) -> pd.DataFrame:
-        assert not os.path.isfile(csv_out_name)
+    def evaluate(
+        self,
+        retriever: BaseHtmlRetriever,
+        rephrased_dataset: pd.DataFrame,
+        csv_out_name: str,
+    ) -> pd.DataFrame:
+        if os.path.isfile(csv_out_name):
+            raise ValueError(f"{csv_out_name} already exists")
         retriever.driver = self.driver
         retrieved_dataset = rephrased_dataset.copy()
         retrieved_dataset["html_id"] = self._gen_html_ids(rephrased_dataset)
-        retrieved_dataset["recall_retriever"] = pd.Series(dtype='float')
-        retrieved_dataset["precision_retriever"] = pd.Series(dtype='float')
-        retrieved_dataset["source_nodes"] = pd.Series(dtype='str')
-        retrieved_dataset["retrieval_time"] = pd.Series(dtype='float')
+        retrieved_dataset["recall_retriever"] = pd.Series(dtype="float")
+        retrieved_dataset["precision_retriever"] = pd.Series(dtype="float")
+        retrieved_dataset["source_nodes"] = pd.Series(dtype="str")
+        retrieved_dataset["retrieval_time"] = pd.Series(dtype="float")
         try:
             for i, row in tqdm(retrieved_dataset.iterrows()):
                 self._load_html(row["html_id"], self.driver)
                 element = self.driver.driver.find_element(By.XPATH, row["xpath"])
-                ground_truth_outer_html = self.driver.execute_script("return arguments[0].outerHTML", element)
+                ground_truth_outer_html = self.driver.execute_script(
+                    "return arguments[0].outerHTML", element
+                )
                 start_time = time()
                 results = retriever.retrieve_html(QueryBundle(row["retriever_query"]))
                 duration = time() - start_time
                 source_nodes = [node.text for node in results]
                 context_str = "\n".join(source_nodes)
-                recall_retriever, precision_retriever = self._intersection_backend_node_id(ground_truth_outer_html, context_str)
+                (
+                    recall_retriever,
+                    precision_retriever,
+                ) = self._intersection_backend_node_id(
+                    ground_truth_outer_html, context_str
+                )
                 new_row = {
                     "recall_retriever": recall_retriever,
                     "precision_retriever": precision_retriever,
@@ -132,57 +160,74 @@ class RetrieverEvaluator(Evaluator):
         return retrieved_dataset
 
     def compare(self, results: Dict[str, pd.DataFrame]):
-        data = []
-        figure = plt.figure(figsize=(12, 8))
-        subfigures = figure.subfigures(3, 3, width_ratios=[2, 6, 2], height_ratios=[2, 4, 2])
-        for df in results.values():
-            data.append(df[['recall_retriever', 'precision_retriever', 'retrieval_time']].mean(axis=0).to_dict())
-        radar_plot(data, list(results.keys()), invert_range=["retrieval_time"], fig=subfigures[1][1], config={"theta_tick_lbls_brk_lng_wrds": False})
-        return figure
+        fig, axes = plt.subplots(1, 3, figsize=(15, 5))
 
+        df = pd.DataFrame()
+        df["precision"] = [df["precision_retriever"].mean() for df in results.values()]
+        df["recall"] = [df["recall_retriever"].mean() for df in results.values()]
+        df["time"] = [df["retrieval_time"].mean() for df in results.values()]
+        df["name"] = list(results.keys())
+
+        rplot = sns.barplot(data=df, x="name", y="recall", ax=axes[0])
+        rplot.set(xlabel="")
+        pplot = sns.barplot(data=df, x="name", y="precision", ax=axes[1])
+        pplot.set(xlabel="")
+        tplot = sns.barplot(data=df, x="name", y="time", ax=axes[2])
+        tplot.set(xlabel="", ylabel="time (secs)")
+
+        return fig
 
     def _gen_html_ids(self, df: pd.DataFrame):
         def add_ids_to_tags(html_string):
-            soup = BeautifulSoup(html_string, 'html.parser')
+            soup = BeautifulSoup(html_string, "html.parser")
             node_id = 1
             for tag in soup.find_all(True):  # The True argument here finds all tags
-                tag['backend_node_id'] = str(node_id)
+                tag["backend_node_id"] = str(node_id)
                 node_id += 1
             return str(soup)
 
-        return df['html'].apply(lambda x: add_ids_to_tags(x))
+        return df["html"].apply(lambda x: add_ids_to_tags(x))
+
 
 class LLMEvaluator(Evaluator):
     def __init__(self):
         self.driver = SeleniumDriverForEval(get_selenium_driver=init_driver)
 
-    def evaluate(self, navigation_engine: NavigationEngine, retrieved_dataset: pd.DataFrame, csv_out_name: str) -> pd.DataFrame:
-        # assert not os.path.isfile(csv_out_name)
+    def evaluate(
+        self,
+        navigation_engine: NavigationEngine,
+        retrieved_dataset: pd.DataFrame,
+        csv_out_name: str,
+        max_retry: int = 1,
+    ) -> pd.DataFrame:
+        if os.path.isfile(csv_out_name):
+            raise ValueError(f"{csv_out_name} already exists")
         llm_dataset = retrieved_dataset.copy()
-        llm_dataset['recall_llm'] = pd.Series(dtype='float')
-        llm_dataset['precision_llm'] = pd.Series(dtype='float')
-        llm_dataset['execution_error'] = pd.Series(dtype='str')
-        llm_dataset['target_outer_html'] = pd.Series(dtype='str')
-        llm_dataset['generated_code'] = pd.Series(dtype='str')
-        llm_dataset['retry'] = pd.Series(dtype='int')
-        llm_dataset['code_generation_time'] = pd.Series(dtype='float')
+        llm_dataset["recall_llm"] = pd.Series(dtype="float")
+        llm_dataset["precision_llm"] = pd.Series(dtype="float")
+        llm_dataset["execution_error"] = pd.Series(dtype="str")
+        llm_dataset["target_outer_html"] = pd.Series(dtype="str")
+        llm_dataset["generated_code"] = pd.Series(dtype="str")
+        llm_dataset["retry"] = pd.Series(dtype="int")
+        llm_dataset["code_generation_time"] = pd.Series(dtype="float")
         try:
             for i, row in tqdm(llm_dataset.iterrows()):
                 self._load_html(row["html_id"], self.driver)
                 element = self.driver.driver.find_element(By.XPATH, row["xpath"])
-                ground_truth_outer_html = self.driver.execute_script("return arguments[0].outerHTML;", element)
-                decontaminated_context_str = re.sub(r' backend_node_id="\d+"', '', row["source_nodes"])
+                ground_truth_outer_html = self.driver.execute_script(
+                    "return arguments[0].outerHTML;", element
+                )
+                decontaminated_context_str = re.sub(
+                    r' backend_node_id="\d+"', "", row["source_nodes"]
+                )
                 generated_code = target_outer_html = execution_error = ""
                 recall_llm = precision_llm = 0.0
-                for r in range(5):
-                    retry = r
-                    try:
-                        start_time = time()
-                        generated_code = navigation_engine.get_action_from_context(decontaminated_context_str, row['llm_query'])
-                        duration = time() - start_time
-                    except Exception as e:
-                        execution_error = str(e)
-                        continue
+                for retry in range(max_retry):
+                    start_time = time()
+                    generated_code = navigation_engine.get_action_from_context(
+                        decontaminated_context_str, row["llm_query"]
+                    )
+                    duration = time() - start_time
                     try:
                         local_scope = {"driver": self.driver.get_driver()}
                         assignment_code = keep_assignments(generated_code)
@@ -194,20 +239,23 @@ class LLMEvaluator(Evaluator):
                             if type(local_scope[v]) == WebElement:
                                 target_element = local_scope[v]
                                 break
-                        target_outer_html = self.driver.execute_script("return arguments[0].outerHTML;", target_element)
-                        recall_llm, precision_llm = self._intersection_backend_node_id(ground_truth_outer_html, target_outer_html)
+                        target_outer_html = self.driver.execute_script(
+                            "return arguments[0].outerHTML;", target_element
+                        )
+                        recall_llm, precision_llm = self._intersection_backend_node_id(
+                            ground_truth_outer_html, target_outer_html
+                        )
                         break
                     except Exception as e:
                         execution_error = str(e)
-                        raise e
                 new_row = {
-                    'recall_llm': recall_llm,
-                    'precision_llm': precision_llm,
-                    'execution_error': execution_error,
-                    'target_outer_html': target_outer_html,
-                    'generated_code': generated_code,
-                    'retry': retry,
-                    'code_generation_time': duration,
+                    "recall_llm": recall_llm,
+                    "precision_llm": precision_llm,
+                    "execution_error": execution_error,
+                    "target_outer_html": target_outer_html,
+                    "generated_code": generated_code,
+                    "retry": retry,
+                    "code_generation_time": duration,
                 }
                 llm_dataset.loc[i, new_row.keys()] = new_row.values()
         except Exception as e:
@@ -217,10 +265,19 @@ class LLMEvaluator(Evaluator):
         return llm_dataset
 
     def compare(self, results: Dict[str, pd.DataFrame]) -> Figure:
-        figure = plt.figure(figsize=(12, 8))
-        subfigures = figure.subfigures(3, 3, width_ratios=[2, 6, 2], height_ratios=[2, 4, 2])
-        data = []
-        for df in results.values():
-            data.append(df[['recall_llm', 'precision_llm', 'code_generation_time', 'retry']].mean(axis=0).to_dict())
-        radar_plot(data, list(results.keys()), invert_range=["code_generation_time", "retry"], fig=subfigures[1][1], config={"theta_tick_lbls_brk_lng_wrds": False})
-        return figure
+        fig, axes = plt.subplots(1, 3, figsize=(15, 5))
+
+        df = pd.DataFrame()
+        df["precision"] = [df["precision_llm"].mean() for df in results.values()]
+        df["recall"] = [df["recall_llm"].mean() for df in results.values()]
+        df["time"] = [df["code_generation_time"].mean() for df in results.values()]
+        df["name"] = list(results.keys())
+
+        rplot = sns.barplot(data=df, x="name", y="recall", ax=axes[0])
+        rplot.set(xlabel="")
+        pplot = sns.barplot(data=df, x="name", y="precision", ax=axes[1])
+        pplot.set(xlabel="")
+        tplot = sns.barplot(data=df, x="name", y="time", ax=axes[2])
+        tplot.set(xlabel="", ylabel="time (secs)")
+
+        return fig
